backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
groq>=0.11.0
python-dotenv>=1.0.0
trafilatura>=1.12.0
pydantic>=2.9.0
sqlalchemy>=2.0.36
psycopg[binary]>=3.2.3
pyjwt>=2.10.0
